% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams.R
\name{access_exam}
\alias{access_exam}
\title{Access an exam with a password}
\usage{
access_exam(exam_name)
}
\arguments{
\item{exam_name}{Name of the exam you want to access (this will be given
in class, along with the password)}
}
\description{
Access an exam with a password
}
\examples{
access_exam("practice_1")
}
