% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams.R
\name{check_exam_avail}
\alias{check_exam_avail}
\title{Check if an exam is available in the \code{fofpack} package}
\usage{
check_exam_avail(exam_name)
}
\arguments{
\item{exam_name}{Name of the exam to check for, e.g. \code{"exam_1"}}
}
\value{
TRUE invisibly if the exam is available, or FALSE otherwise
}
\description{
Check if an exam is available in the \code{fofpack} package
}
\examples{
check_exam("exam_1")
}
