% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\name{write_answers}
\alias{write_answers}
\title{Write your submitted answers from the assignment to a text file}
\usage{
write_answers(week, where = getwd())
}
\arguments{
\item{week}{What week's assignment do you want your answers for? This should
be numeric.}

\item{where}{Where do you want it? That is, in which directory on your computer
do you want to write the text file. Defaults to your working directory.}
}
\value{
The path to a text file with answers
}
\description{
Get a text file of your answers to the assignment that you can upload to
Canvas.
}
\examples{
write_answers(3)
}
