% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fofpack-package.R
\docType{package}
\name{fofpack-package}
\alias{fofpack}
\alias{fofpack-package}
\title{fofpack: Forest of the Future Course Package}
\description{
Course material for FIU course 'Forest of the Future'
}
\author{
\strong{Maintainer}: Russell Dinnage \email{rdinnage@fiu.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
